+++ azure/Kqlmagic/cache_client.py
@@ -34,18 +34,18 @@ class CacheClient(object):
 
 
     @property
-    def data_source(self):
+    def data_source(self) -> str:
         return self.files_folder
 
 
-    def _get_query_hash_filename(self, query):
+    def _get_query_hash_filename(self, query: str) -> str:
         lines = [l.replace("\r", "").replace("\t", " ").strip() for l in query.split("\n")]
         q_lines = []
         for line in lines:
             if not line.startswith("//"):
                 idx = line.find(" //")
                 q_lines.append(line[: idx if idx >= 0 else len(line)])
-        return "q_" + hashlib.sha1(bytes("".join(q_lines), "utf-8")).hexdigest() + ".json"
+        return f"q_{hashlib.sha1(bytes(''.join(q_lines), 'utf-8')).hexdigest()}.json"
 
 
     def _get_file_path(self, query, database_at_cluster, cache_folder):