+++ azure/Kqlmagic/kusto_client.py
@@ -225,12 +225,16 @@ class Kusto_Client(object):
         # print("response text", response.text)
 
         if response.status_code != requests.codes.ok:  # pylint: disable=E1101
-            raise KqlError([response.text], response)
+            raise KqlError(response.text, response)
 
         kql_response = KqlQueryResponse(response.json(), endpoint_version)
 
         if kql_response.has_exceptions() and not accept_partial_results:
-            raise KqlError(kql_response.get_exceptions(), response, kql_response)
+            try:
+                error_message = json.dumps(kql_response.get_exceptions())
+            except:
+                error_message = str(kql_response.get_exceptions())
+            raise KqlError(error_message, response, kql_response)
 
         return kql_response